refactor: move status into cli
