+++ packages/cli/test/unit/commands/status.unit.test.ts
